% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.entropyProfile.R
\name{as.data.frame.entropyProfile}
\alias{as.data.frame.entropyProfile}
\title{Coerce \code{entropyProfile} to a Data Frame}
\usage{
\method{as.data.frame}{entropyProfile}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{entropyProfile}.}

\item{row.names}{Please see \code{\link{as.data.frame}}.}

\item{optional}{Please see \code{\link{as.data.frame}}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A \code{data frame} with tabular information on an entropy profile.
This information includes the name of the proteins presenting
mutations, the corresponding genomic positions, and the resulting
entropies in the metagenome.
}
\description{
Function to extract summary information from an \code{entropyProfile}
object. This function is internally used for plotting.
}
