% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMap_entropyProfiles.R
\name{heatmap_entropyProfiles}
\alias{heatmap_entropyProfiles}
\title{Graphically compare entropy profiles}
\usage{
heatmap_entropyProfiles(..., keepInvariant = TRUE, showAllPeptides = TRUE)
}
\arguments{
\item{...}{Unquoted \code{entropyProfile} data objects to be compared.}

\item{keepInvariant}{Logical; whether to retain positions where entropy is
zero. Defaults to \code{FALSE}.}

\item{showAllPeptides}{Logical; whether the heatmap should show all peptides
encoded by the virus, regardless of whether they have undergone mutations.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the heatmap.
}
\description{
This function prints heatmaps of multiple entropy profiles
(\code{entropyProfile} objects).
}
\examples{
firstWave <- wWater[ wWater$wave == "first", ]
thirdWave <- wWater[ wWater$wave == "third", ]
# Filter out minor variants whose genotype matches that of the reference genome
firstWave <- firstWave[ firstWave$alt_aa_freq <= 0.97, ]
thirdWave <- thirdWave[ thirdWave$alt_aa_freq <= 0.97, ]
ancestral <- getEntropySignature(firstWave)
omicron <-  getEntropySignature(thirdWave)

# High entropies sustained over time in the spike protein, and signals
# compatible with pervasive negative selection in Omicron sublineages:
heatmap_entropyProfiles(ancestral, omicron)

}
