% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn908947.3.R
\docType{data}
\name{mn908947.3}
\alias{mn908947.3}
\title{CDS topology and length of Wuhan-Hu-1 reference strain}
\format{
An object of class \code{list} of length 2.
}
\source{
Nucleotide [Internet]. Bethesda (MD): National Library of Medicine (US),
National Center for Biotechnology Information; [1988] – . Accession No.
MN908947.3, Severe acute respiratory syndrome coronavirus 2 isolate Wuhan-Hu-1, complete genome.
Available from: \url{https://www.ncbi.nlm.nih.gov/nuccore/MN908947.3}.
}
\usage{
mn908947.3
}
\description{
This type of data can be obtained from \code{.gff} (General Feature Format)
files using applications such as the \code{rtracklayer} package, or manually
from the corresponding entry in the GenBank database.
}
\keyword{datasets}
