% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.entropyProfile.R
\name{summary.entropyProfile}
\alias{summary.entropyProfile}
\title{Summarize entropy profile}
\usage{
\method{summary}{entropyProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{entropyProfile}.}

\item{...}{Other parameters passed to the function.}
}
\value{
An object of class \code{c("profileSummary", "list")} summarizing
an entropy profile. Intended to be displayed via
\code{print.profileSummary}.
}
\description{
Prints a report about an entropy profile (an object of class
"entropyProfile").
}
