% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wWater.R
\docType{data}
\name{wWater}
\alias{wWater}
\title{Data from first and third COVID-19 waves in Trelew \url{http://tools.wmflabs.org/geohack/geohack.php?language=es&pagename=Trelew&params=-43.253333333333_N_-65.309444444444_E_type:city}}
\format{
An object of class \code{data.frame} with 148 rows and 10 columns.
}
\source{
Manrique, Julieta Marina, and Leandro Roberto Jones. 2026. Genomic
surveillance of SARS-CoV-2: a resource-efficient wastewater-based
workflow. Molecular Biology Reports 53 : 995.
}
\usage{
wWater
}
\description{
SNVs inferred from Illumina (2 x 150) sequences from pooled ultra-pure virus
concentrates representative of the 1st and 3rd COVID-19 waves in Trelew.
Reads were mapped against the Wuhan-Wu-1 reference genome (MN908947.3) by
\emph{bwa}, and variants were called with \emph{iVar} with a 3\% frequency
cutoff for minor variants.
First wave cases were caused by ancestral strains whereas third wave cases
were mainly due to highly human-adapted Omicron sublineages.
}
\keyword{datasets}
