% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosEntropy.R
\name{getPosEntropy}
\alias{getPosEntropy}
\title{Calculates aminoacidic entropy resulting from SNVs present in a specific locus}
\usage{
getPosEntropy(variantPosition)
}
\arguments{
\item{variantPosition}{A list, created  and passed to the function by
\code{fillPosition}. It contains information on the
frequencies of each amino acid category observed in a
virome, as a result of mutations at a given genomic
position.}
}
\value{
A \code{numeric} value, corresponding to the entropy associated to
the SNVs present at a specific locus.
}
\description{
The function is used internally by \code{\link{getEntropySignature}}.
It calculates Shannon entropy from categories and frequencies passed to the
function.
}
\seealso{
\code{\link{getEntropySignature}}.
}
\keyword{internal}
