% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.profileSummary.R
\name{print.profileSummary}
\alias{print.profileSummary}
\title{Print method for  \code{profileSummary} objects}
\usage{
\method{print}{profileSummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{profileSummary} created by
\code{\link{summary.entropyProfile}}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
Invisibly returns \code{NULL}. This function is used for its side
effect.
}
\description{
This function formats and prints compact entropy profile summaries
(\code{profileSummary} objects), on the console.
}
