% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tydyMutations.R
\name{print.tidyMutations}
\alias{print.tidyMutations}
\title{Print method for  \code{tidyMutations} objects}
\usage{
\method{print}{tidyMutations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidyMutations} created by
\code{\link{showMutations}}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
Invisibly returns \code{NULL}. Called for side effect.
}
\description{
This function formats and prints compact mutation summaries
(\code{tidyMutations} objects) created by showMutations(), on the console.
}
